fix nextjs page exports and optimize nft image
