+++ packages/nextjs/utils/protocolData.ts
@@ -1,0 +1,38 @@
+export const protocols = [
+  "Base",
+  "Bitcoin",
+  "Celo",
+  "Ethereum",
+  "Gitcoin",
+  "Kleros",
+  "Monero",
+  "Optimism",
+  "UMA",
+  "Uniswap",
+];
+
+export const descriptions = [
+  ", the agile mage who channels Coinbase’s strength into a fast, low-cost Layer-2 stage built on the OP Stack, inviting builders to conjure dApps with ease.",
+  ", the stoic grandmaster who introduced the world to peer-to-peer electronic cash, forging trustless value with cryptographic proof instead of intermediaries.",
+  ", the nimble enchantress who puts mobile first, casting carbon-negative spells to make ultra-light, low-fee payments accessible to everyone.",
+  ", the versatile archmage who transforms code into unstoppable smart contracts, powering a universe of decentralized applications on his global ledger.",
+  ", the generous benefactor whose quadratic-funding magic rallies communities to finance open-source public goods with fairness and impact.",
+  ", the impartial adjudicator whose decentralized jury spells deliver fast, affordable verdicts for Web3 disputes across realms.",
+  ", the shadow-weaver who cloaks every sender, receiver, and amount in ring signatures and stealth addresses, championing default privacy for all.",
+  ", the optimistic sorceress scaling Ethereum with rollup magic, granting users cheaper, speedier transactions while preserving mainnet security.",
+  ", the truth-seeker who wields an Optimistic Oracle to anchor any verifiable fact on-chain, opening markets for limitless synthetic assets.",
+  ", the liquidity alchemist whose automated market-maker formula turns pooled tokens into constant, permissionless swaps for traders worldwide.",
+];
+
+export const websites = [
+  "https://base.org",
+  "https://bitcoin.org",
+  "https://celo.org",
+  "https://ethereum.org",
+  "https://gitcoin.co",
+  "https://kleros.io",
+  "https://www.getmonero.org",
+  "https://www.optimism.io",
+  "https://uma.xyz",
+  "https://uniswap.org",
+];
